pyproject.toml
[project]
name = "autoinvoice"
version = "0.1.0"
description = "Google Sheets payroll data to MoneyForward Cloud Invoice automation"
requires-python = ">=3.9"
dependencies = [
    "click>=8.0",
    "gspread>=6.0",
    "google-auth>=2.0",
    "google-api-python-client>=2.0",
    "requests>=2.28",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
autoinvoice = "autoinvoice.cli:main"

[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
